Cleaning up namespace and folder structure for System.UnitTesting
