+++ src/System.ComponentModel.Composition/tests/System.ComponentModel.Composition.Tests.csproj
@@ -156,19 +156,6 @@
     <Compile Include="System\ComponentModel\Composition\ScopedCompositionServicetests.cs" />
     <Compile Include="System\ComponentModel\Composition\ScopeExportFactoryTests.cs" />
     <Compile Include="System\ComponentModel\Composition\ScopingHelpers.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\CompositionAssert.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\ConstraintAssert.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\EnumerableAssert.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\ErrorId.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\ExceptionAssert.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\ExpectationCollectionOfI.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\ExpectationCollectionOfIO.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\ExpectationOfI.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\ExpectationOfIO.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\ExportsAssert.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\ExtendedAssert.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\RetryMode.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\TestServices.cs" />
     <Compile Include="System\Globalization\CurrentCultureContext.cs" />
     <Compile Include="System\Integration\AdaptingCollectionTests.cs" />
     <Compile Include="System\Integration\CatalogFilteringTests.cs" />
@@ -184,6 +171,21 @@
     <Compile Include="System\Integration\RequiredCreationPolicyTests.cs" />
     <Compile Include="System\ReferenceTracker.cs" />
     <Compile Include="System\TypeExtensions.cs" />
+    <Compile Include="System\UnitTesting\AssertExtensions.cs" />
+    <Compile Include="System\UnitTesting\CompositionAssert.cs" />
+    <Compile Include="System\UnitTesting\ConstraintAssert.cs" />
+    <Compile Include="System\UnitTesting\EnumerableAssert.cs" />
+    <Compile Include="System\UnitTesting\EqualityExtensions.cs" />
+    <Compile Include="System\UnitTesting\ErrorId.cs" />
+    <Compile Include="System\UnitTesting\ExceptionAssert.cs" />
+    <Compile Include="System\UnitTesting\ExpectationCollectionOfI.cs" />
+    <Compile Include="System\UnitTesting\ExpectationCollectionOfIO.cs" />
+    <Compile Include="System\UnitTesting\ExpectationOfI.cs" />
+    <Compile Include="System\UnitTesting\ExpectationOfIO.cs" />
+    <Compile Include="System\UnitTesting\ExportsAssert.cs" />
+    <Compile Include="System\UnitTesting\ExtendedAssert.cs" />
+    <Compile Include="System\UnitTesting\RetryMode.cs" />
+    <Compile Include="System\UnitTesting\TestServices.cs" />
     <Compile Include="TestAssembly.cs" />
     <Compile Include="TransparentTestCase.cs" />
   </ItemGroup>
