+++ src/System.ComponentModel.Composition/tests/System/ComponentModel/Composition/ComposablePartExtensibilityTests.cs
@@ -6,7 +6,6 @@ using System.Collections.Generic;
 using System.ComponentModel.Composition.Factories;
 using System.ComponentModel.Composition.Hosting;
 using System.ComponentModel.Composition.Primitives;
-using System.ComponentModel.Composition.UnitTesting;
 using System.Linq;
 using System.Reflection;
 using System.UnitTesting;
