+++ src/System.ComponentModel.Composition/tests/System/ComponentModel/Composition/CompositionResultTest.cs
@@ -4,7 +4,6 @@
 
 using System.Collections.Generic;
 using System.ComponentModel.Composition.Factories;
-using System.ComponentModel.Composition.UnitTesting;
 using System.Linq;
 using System.UnitTesting;
 using Xunit;
