+++ src/System.ComponentModel.Composition/tests/System/ComponentModel/Composition/ContractBasedImportDefinitionTests.cs
@@ -5,7 +5,6 @@
 using System.Collections.Generic;
 using System.ComponentModel.Composition.Hosting;
 using System.ComponentModel.Composition.Primitives;
-using System.ComponentModel.Composition.UnitTesting;
 using System.Linq;
 using System.UnitTesting;
 using Xunit;
