+++ src/System.ComponentModel.Composition/tests/System/ComponentModel/Composition/ExportCollectionTests.cs
@@ -7,7 +7,6 @@ using System.Collections.ObjectModel;
 using System.ComponentModel.Composition.Factories;
 using System.ComponentModel.Composition.Hosting;
 using System.ComponentModel.Composition.Primitives;
-using System.ComponentModel.Composition.UnitTesting;
 using System.Linq;
 using System.UnitTesting;
 using Xunit;
