+++ src/System.ComponentModel.Composition/tests/System/Integration/CompositionContainerAttributedModelCycleTests.cs
@@ -6,8 +6,8 @@ using System;
 using System.ComponentModel.Composition;
 using System.ComponentModel.Composition.Factories;
 using System.ComponentModel.Composition.Primitives;
-using System.ComponentModel.Composition.UnitTesting;
 using System.Linq;
+using System.UnitTesting;
 using Xunit;
 
 namespace Tests.Integration
