+++ src/System.ComponentModel.Composition/tests/System/Integration/LifetimeTests.cs
@@ -7,7 +7,6 @@ using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.ComponentModel.Composition.Factories;
 using System.ComponentModel.Composition.Hosting;
-using System.ComponentModel.Composition.UnitTesting;
 using System.Linq;
 using System.Reflection;
 using System.UnitTesting;
