+++ src/System.ComponentModel.Composition/tests/System/UnitTesting/CompositionAssert.cs
@@ -2,11 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.ComponentModel.Composition;
 using System.ComponentModel.Composition.Primitives;
 using System.UnitTesting;
 using Xunit;
 
-namespace System.ComponentModel.Composition.UnitTesting
+namespace System.UnitTesting
 {
     public static class CompositionAssert
     {
