+++ src/System.ComponentModel.Composition/tests/System/UnitTesting/ExceptionAssert.cs
