+++ src/System.ComponentModel.Composition/tests/System/UnitTesting/ExportsAssert.cs
@@ -8,7 +8,7 @@ using System.ComponentModel.Composition.Primitives;
 using System.Linq;
 using Xunit;
 
-namespace System.ComponentModel.Composition.UnitTesting
+namespace System.UnitTesting
 {
     public static class ExportsAssert
     {
